checkUserIsAdmin

+++ pecado-ims/pecado-ims-web/src/main/java/me/batizhao/ims/service/impl/UserServiceImpl.java
@@ -7,6 +7,7 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import me.batizhao.common.core.exception.NotFoundException;
+import me.batizhao.common.core.exception.PecadoException;
 import me.batizhao.ims.api.domain.Menu;
 import me.batizhao.ims.api.domain.Role;
 import me.batizhao.ims.api.domain.User;
@@ -108,10 +109,10 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
 
         BCryptPasswordEncoder bcryptPasswordEncoder = new BCryptPasswordEncoder();
         if (!bcryptPasswordEncoder.matches(oldPassword, user.getPassword()))
-            throw new RuntimeException("旧密码不正确");
+            throw new PecadoException("旧密码不正确");
 
         if (bcryptPasswordEncoder.matches(newPassword, user.getPassword()))
-            throw new RuntimeException("新旧密码相同");
+            throw new PecadoException("新旧密码相同");
 
         String hashPass = bcryptPasswordEncoder.encode(newPassword);
         user.setPassword(hashPass);
@@ -122,11 +123,18 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     @Override
     @Transactional
     public Boolean deleteByIds(List<Long> ids) {
-        this.removeByIds(ids);
         ids.forEach(i -> {
+            checkUserIsAdmin(i);
             userRoleService.remove(Wrappers.<UserRole>lambdaQuery().eq(UserRole::getUserId, i));
         });
+        this.removeByIds(ids);
         return true;
+    }
+
+    private void checkUserIsAdmin(Long id) {
+        if (id.equals(1L)) {
+            throw new PecadoException("管理员不允许操作！");
+        }
     }
 
     @Override
